Querying components from OnDestroy callback now properly returns only non-destroyed components (fixes a crash on exit for Animation component) Animation proxies are now properly visible on the animation thread Final skeleton poses are now properly visible on the core thread Checking if a Renderable has an animation now returns a valid answer C# Animation OnEnable/OnDisable callbacks now trigger properly Pause/unpause animation when exiting/entering play-mode in editor
Former-commit-id: 6869dbcfbc42243adb5b0218108aa32af26cb584

+++ Source/BansheeEngine/Include/BsRenderable.h
@@ -168,7 +168,7 @@ namespace BansheeEngine
 		UINT32 getRendererId() const { return mRendererId; }
 
 		/** Checks is the mesh geometry rendered by this renderable animated using skeleton or blend shape animation. */
-		bool isAnimated() const { return mAnimationId != (UINT32)-1; }
+		bool isAnimated() const { return mAnimationId != (UINT64)-1; }
 
 		/** Returns the identifier of the animation, if this object is animated using skeleton or blend shape animation. */
 		UINT64 getAnimationId() const { return mAnimationId; }
